Adjust SoilIce Calculations
Tweaked permafrost so that if permafrost is found in September, the code no longer checks for permafrost in winter. It assumes there is permafrost, and will set frostDepth to leakage frost depth + rooting depth for all following months until permafrost is potentially turned off next September

ActiveLayerDepth is now set to zero in non-summer months.

When SoilIceDepth is set to false and activeLayerDepth becomes undefined, set it to 999.